move some high order mesh tests back into unit tests
